refactor: split trade test game builders (wave 12 slice E, #4240)
Extract Game factories from trade_screen_test_support.dart into
trade_screen_test_game_builders.dart so the support module stays under
the wave-12 ≤350 physical-line target while preserving re-exports.

Refs #4240
