+++ tool/check_app_test_no_duplicate_scaffolding.dart
@@ -180,7 +180,9 @@ int runCheckAppTestNoDuplicateScaffolding(
     'pumpDialogs320At from app/test/dialogs_320dp_min_viewport_support.dart. '
     'Widgetbook: use findWidgetbookUseCase from '
     'app/test/widgetbook_test_harness.dart. '
-    'Trade: use buildTradeTestGame / pumpTradeScreen* from '
+    'Trade: use buildTradeTestGame from '
+    'app/test/trade_screen_test_game_builders.dart (re-exported from '
+    'trade_screen_test_support.dart) and pumpTradeScreen* from '
     'app/test/trade_screen_test_support.dart. '
     'Units-panel: use shared factories in '
     'civilian_units_panel_test_support.dart / '
